requirements.txt
tensorflow-gpu==2.6.0
numpy>=1.19.2
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
opencv-python>=4.5.3
captum>=0.4.0
torch>=1.9.0  # Required for Captum
pillow>=8.3.2
h5py>=3.1.0
tqdm>=4.62.3 